requirements.txt
# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP client for API calls
httpx==0.25.2

# Data validation and serialization
pydantic==2.5.0

# File upload support (if needed later)
python-multipart==0.0.6

# Environment variable management
python-dotenv==1.0.0

# Optional: For better development experience
# pytest==7.4.3
# pytest-asyncio==0.21.1

